Tooling: download remote closure to compare
It happens we GC the store locally or a given host has been deployed
from elsewhere and we don't have the whole closure for a remote system
and therefore the script comparing what's changed in the meantime wasn't
always able to check.

This patch make sure we have it locally first. Of course it's not ideal
for the local store size.